requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
aiohttp>=3.9.0
pillow>=10.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
pydantic>=2.5.0
python-multipart>=0.0.6
opencv-python>=4.8.0
psutil>=5.9.0
